requirements.txt
torch
torchvision
sklearn
matplotlib
numpy
pandas
Pillow
